backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
pyjwt==2.8.1
bcrypt==4.1.1
pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
python-docx==0.8.11
reportlab==4.0.7
aiofiles==23.2.1
email-validator==2.1.0
